bugfix: set AllowCreate in NameIDPolicy to true by default
This was previously the case. However, when we introduced the new way to configure NameIDPolicy by using arrays, the default for old-style string configurations changed so AllowCreate was then unset. When unset, the SAML standard sets its default to false, effectively creating interoperability issues when uses upgraded SimpleSAMLphp without changing the configuration.

This resolves #1230.